+++ modules/ALU/tb.cpp
@@ -5,20 +5,20 @@
 #include <time.h>
 #include <cmath>
 #include <iostream>
-#include <VRAM__Syms.h>
+#include <VALU__Syms.h>
 #include <assert.h>
 
 using namespace std;
 
-VRAM *dut = new VRAM;
+VALU *dut = new VALU;
 
 #define IS_SIM_TIME_IN_RST(sim_time) (sim_time >= 3 && sim_time < 6)
 #define MAX_SIM_TIME 300
 #define VERIF_START_TIME 7
 #define MAX_STAGE 100
-#define myexit(condition, content)    \
-    {                                 \
-        assert(condition && content); \
+#define myexit(condition, content)   \
+    {                                \
+        assert(condition &&content); \
     }
 
 vluint64_t sim_time = 0;
@@ -47,7 +47,7 @@ vluint8_t combinational_logic_update = COMBINATIONAL_LOGIC_EVAL_EN;
             LATCH_MANAGEMENT_SELECTOR_TO_LATCH(lm);                                                        \
             LATCH_MANAGEMENT_LATCH_ASSIGN(lm, lm.after_latch_state);                                       \
             statement1                                                                                     \
-        }                                                                                                  \
+        }                                                                                                 \
         else                                                                                               \
         {                                                                                                  \
             LATCH_MANAGEMENT_SELECTOR_TO_AFTER_LATCH(lm);                                                  \
@@ -68,31 +68,23 @@ typedef struct
     uint64_t *selector;
 } latch_management;
 
-class RAMInTx
+class ALUInTx
 {
 public:
     /* TODO BEGIN 1 */
-    uint8_t clk,
-        rst_n,
-
-        write_en,
-        write_addr,
-        write_data,
-
-        read_en,
-        read_addr;
+    myexit(0, "Delete me first before filling this TODO")
     /* TODO END 1 */
 };
 
-class RAMOutTx
+class ALUOutTx
 {
 public:
     /* TODO BEGIN 2 */
-    uint8_t read_data;
+    myexit(0, "Delete me first before filling this TODO")
     /* TODO END 2 */
 };
 
-class RAMInternalTx
+class ALUInternalTx
 {
 public:
     /* TODO BEGIN 2 */
@@ -100,68 +92,41 @@ public:
     /* TODO END 2 */
 };
 
-RAMInTx in_tx_ref;
-RAMOutTx out_tx_ref;
-// RAMInternalTx internal_tx_ref;
+// ALUInTx in_tx_ref;
+// ALUOutTx out_tx_ref;
+// ALUInternalTx internal_tx_ref;
 
-class RAMScb
+class ALUScb
 {
 private:
-    std::deque<RAMInTx *> in_q;
+    std::deque<ALUInTx *> in_q;
 
 public:
     // Input interface monitor port
-    void writeIn(RAMInTx *tx)
+    void writeIn(ALUInTx *tx)
     {
         // Push the received transaction item into a queue for later
         in_q.push_back(tx);
     }
 
     // Output interface monitor port
-    void writeOut(RAMOutTx *tx)
+    void writeOut(ALUOutTx *tx)
     {
         // We should never get any data from the output interface
         // before an input gets driven to the input interface
         if (in_q.empty())
         {
-            std::cout << "Fatal Error in RAMScb: empty RAMInTx queue" << std::endl;
+            std::cout << "Fatal Error in ALUScb: empty ALUInTx queue" << std::endl;
             exit(1);
         }
 
         // Grab the transaction item from the front of the input item queue
-        RAMInTx *in;
+        ALUInTx *in;
         in = in_q.front();
         in_q.pop_front();
 
         /* TODO BEGIN 3 */
-        if (!in->rst_n)
-        {
-            if (!(tx->read_data == 0x0))
-            {
-                printf("\r\n# TODO 3 Failed at simtime %ld", sim_time);
-                printf("\r\n# TODO 3 INPUT TRACE: in->read_addr = 0x%x, in->read_en = 0x%x, in->rst_n = 0x%x, in->write_addr = 0x%x, in->write_data = 0x%x, in->write_en = 0x%x", in->read_addr, in->read_en, in->rst_n, in->write_addr, in->write_data, in->write_en);
-                printf("\r\n# TODO 3 OUTPUT TRACE: tx->read_data = 0x%x", tx->read_data);
-                printf("\r\n");
-                fflush(stdout);
-
-                myexit(tx->read_data == 0x0, "TODO 3 Failed: Reset logic result of the Verilog module is incorrect")
-            }
-        }
-        else if (in->read_en)
-        {
-            if (!(tx->read_data == out_tx_ref.read_data))
-            {
-                printf("\r\n# TODO 3 Failed at simtime %ld", sim_time);
-                printf("\r\n# TODO 3 INPUT TRACE: in->read_addr = 0x%x, in->read_en = 0x%x, in->rst_n = 0x%x, in->write_addr = 0x%x, in->write_data = 0x%x, in->write_en = 0x%x", in->read_addr, in->read_en, in->rst_n, in->write_addr, in->write_data, in->write_en);
-                printf("\r\n# TODO 3 OUTPUT TRACE: tx->read_data = 0x%x", tx->read_data);
-                printf("\r\n# TODO 3 REFERENCE OUTPUT TRACE: out_tx_ref.read_data = 0x%x", out_tx_ref.read_data);
-                printf("\r\n");
-                fflush(stdout);
-
-                myexit(tx->read_data == out_tx_ref.read_data, "TODO 3 Failed: Memory logic result of the Verilog module is incorrect")
-            }
-        }
-
+        myexit(0, "Delete me first before filling this TODO")
         /* TODO END 3 */
 
         delete in;
@@ -169,146 +134,93 @@ public:
     }
 };
 
-class RAMInDrv
+class ALUInDrv
 {
 private:
-    VRAM *dut;
+    VALU *dut;
 
 public:
-    RAMInDrv(VRAM *dut)
+    ALUInDrv(VALU *dut)
     {
         this->dut = dut;
     }
 
-    void drive(RAMInTx *tx)
+    void drive(ALUInTx *tx)
     {
         /* TODO BEGIN 4 */
+        myexit(0, "Delete me first before filling this TODO")
         if (tx != NULL)
         {
-
-            dut->read_addr = tx->read_addr;
-            dut->read_en = tx->read_en;
-            dut->write_addr = tx->write_addr;
-            dut->write_data = tx->write_data;
-            dut->write_en = tx->write_en;
-
             if (COMBINATIONAL_LOGIC_EVAL_EN)
                 dut->eval(); // combinational update
-
-            dut->rst_n = tx->rst_n;
             delete tx;
         }
         /* TODO END 4 */
-
+        
         dut->clk ^= IS_SEQUENTIAL_LOGIC_UPDATE(combinational_logic_update);
         dut->eval(); // sequential update
     }
 };
 
-class RAMInMon
+class ALUInMon
 {
 private:
-    VRAM *dut;
-    RAMScb *scb;
+    VALU *dut;
+    ALUScb *scb;
 
 public:
-    RAMInMon(VRAM *dut, RAMScb *scb)
+    ALUInMon(VALU *dut, ALUScb *scb)
     {
         this->dut = dut;
         this->scb = scb;
     }
     void monitor()
     {
-        RAMInTx *tx = new RAMInTx();
+        ALUInTx *tx = new ALUInTx();
 
         /* TODO BEGIN 5 */
-        tx->read_addr = dut->read_addr;
-        tx->read_en = dut->read_en;
-        tx->write_addr = dut->write_addr;
-        tx->write_data = dut->write_data;
-        tx->write_en = dut->write_en;
-        tx->rst_n = dut->rst_n;
+        myexit(0, "Delete me first before filling this TODO")
         /* TODO END 5 */
 
         scb->writeIn(tx);
     }
 };
 
-class RAMOutMon
+class ALUOutMon
 {
 private:
-    VRAM *dut;
-    RAMScb *scb;
+    VALU *dut;
+    ALUScb *scb;
 
 public:
-    RAMOutMon(VRAM *dut, RAMScb *scb)
+    ALUOutMon(VALU *dut, ALUScb *scb)
     {
         this->dut = dut;
         this->scb = scb;
     }
     void monitor()
     {
-        RAMOutTx *tx = new RAMOutTx();
-
+        ALUOutTx *tx = new ALUOutTx();
+        
         /* TODO BEGIN 6 */
-        tx->read_data = dut->read_data;
+        myexit(0, "Delete me first before filling this TODO")
         /* TODO END 6 */
 
         scb->writeOut(tx);
     }
 };
 
-RAMInTx *rndAluInTx()
+ALUInTx *rndAluInTx()
 {
-    RAMInTx *tx = new RAMInTx();
+    ALUInTx *tx = new ALUInTx();
     /* TODO BEGIN 7 */
-    uint8_t tx_data_gen_time_increase = IS_SEQUENTIAL_LOGIC_EVAL(!dut->clk, combinational_logic_update);
+    uint8_t tx_data_gen_time_increase = IS_COMBINATIONAL_LOGIC_CONDITION_EVAL(combinational_logic_update, !dut->CLK);
     if (IS_SIM_TIME_IN_RST(sim_time))
-        tx->rst_n = 0;
-
+        tx->rst = 0;
+    
+    myexit(0, "Delete me first before filling this TODO")
     else if (sim_time >= VERIF_START_TIME)
     {
-        if (tx_data_gen_time_increase)
-            switch (tx_data_gen_time)
-            {
-            case 0:
-                in_tx_ref.rst_n = 1;
-                in_tx_ref.write_en = 0;
-                in_tx_ref.read_en = 0;
-                in_tx_ref.read_addr = in_tx_ref.write_addr = 0x7;
-                in_tx_ref.write_data = 0;
-                break;
-            case 1:
-                in_tx_ref.read_addr++;
-                in_tx_ref.write_addr++;
-                in_tx_ref.read_addr &= 0x7;
-                in_tx_ref.write_addr &= 0x7;
-
-                in_tx_ref.write_en = 1;
-                in_tx_ref.read_en = 0;
-
-                in_tx_ref.write_data = out_tx_ref.read_data = rand() & 0x3f;
-                break;
-            case 2:
-                in_tx_ref.write_en = 0;
-                in_tx_ref.read_en = 1;
-                break;
-            case 3:
-                tx_data_gen_time_increase = 0;
-                tx_data_gen_time = 1;
-                break;
-            default:
-
-                break;
-            }
-
-        tx->read_addr = in_tx_ref.read_addr;
-        tx->read_en = in_tx_ref.read_en;
-        tx->rst_n = in_tx_ref.rst_n;
-        tx->write_addr = in_tx_ref.write_addr;
-        tx->write_data = in_tx_ref.write_data;
-        tx->write_en = in_tx_ref.write_en;
-
         tx_data_gen_time += tx_data_gen_time_increase;
         tx_data_gen_time %= MAX_STAGE;
     }
@@ -331,15 +243,16 @@ int main(int argc, char **argv)
     dut->trace(m_trace, 5);
     m_trace->open("waveform.vcd");
 
-    RAMInTx *tx;
+    ALUInTx *tx;
 
     // Here we create the driver, scoreboard, input and output monitor blocks
-    RAMInDrv *drv = new RAMInDrv(dut);
-    RAMScb *scb = new RAMScb();
-    RAMInMon *inMon = new RAMInMon(dut, scb);
-    RAMOutMon *outMon = new RAMOutMon(dut, scb);
+    ALUInDrv *drv = new ALUInDrv(dut);
+    ALUScb *scb = new ALUScb();
+    ALUInMon *inMon = new ALUInMon(dut, scb);
+    ALUOutMon *outMon = new ALUOutMon(dut, scb);
 
     /* TODO BEGIN 8 */
+    myexit(0, "Delete me first before filling this TODO")
     while (sim_time < MAX_SIM_TIME)
     {