+++ modules/RAM/verified_RAM.v
@@ -6,22 +6,22 @@ module RAM # (
 	input rst_n,
 	
 	input write_en,
-	input [WIDTH-1:0]write_addr,
-	input [DEPTH-1:0]write_data,
+	input [$clog2(DEPTH)-1:0]write_addr,
+	input [WIDTH-1:0]write_data,
 	
 	input read_en,
-	input [WIDTH-1:0]read_addr,
-	output reg [DEPTH-1:0]read_data
+	input [$clog2(DEPTH)-1:0]read_addr,
+	output reg [WIDTH-1:0]read_data
 );
     
     //defination
-    reg [DEPTH-1:0] internal_RAM [2**WIDTH-1:0];
+    reg [WIDTH-1:0] internal_RAM [DEPTH-1:0];
 
     //output 
     integer i;
     always@(posedge clk or negedge rst_n)begin
         if(!rst_n) begin
-               for(i = 0; i < 2**WIDTH; i = i + 1) begin
+               for(i = 0; i < DEPTH; i = i + 1) begin
                    internal_RAM[i] <= 'd0;
                end
         end